vim: Move autosource of config into ftplugin
Instead of only sourcing the vim config when vimrc or $MYVIMRC (xdg.vim)
is edited, reload always when saving a vim file.
This is a test. But I think it makes sense since I never edit a vim file
that is not sourced by my config in any way. Thus reloading should be
always appropriate.