git: Enable jumps between different files in diff
Jump between file sections and commits by using n and N. Taken from:
https://github.com/so-fancy/diff-so-fancy/blob/eef379ee43f86/pro-tips.md

Remove FX flags since they are incompatible with the pattern search:
When showing a small change that fits on less than the screen height,
the rest of the screen is filled with ~ lines instead of quitting the
pager. This behaviour is annoying and thus the flags are deactivated.
Since git adds them per default they have to be explicitly disabled.