vim: Load other always-on plugins too
And move the plugin specific keybindings into 40-keybindings.vim.
`packloadall` sources the plugins directly in contrast to `packadd`.
Because of that we can access the `g:loaded_*` variables and do not have
to specify our own keybindings for vim-surround as now the user bindings
are loaded later and overwrite possible clashes.