vim: Move vimtex settings into tex ftplugin
And load vimtex automatically. I discarded the autocommand as vimtex
only loads a small ftdetect script for tex and the rest of the plugin is
made up of ftplugins, indent files, etc. by itself.
The overhead of the autocmd feels bigger than just loading the small
component and letting vim doing the rest.
(But this is in now way tested/benchmarked. That is just how I feel and
it makes it a tad cleaner)