xmodmap: Change layout and disable xcape
Change layout so that instead of rebinding caps_lock, escape and ctrl_r
in a circle, the caps_lock key now just emulates escape. Caps_lock can now be
accessed by pressing both Shifts together.

Because of that xcape is disabled for now too. Not sure if I will keep
it this way.
Caps_lock is now triggered on every shift for a very small time too,
that is why the caps_lock binding is removed from the i3 config.