zathura: Disable incremental-search for now
Disable the option as the search stops taking any input as soon as there
are no search results left. When typing a word fast but making a typo
this is very annoying. Zathura will stop then the search in the middle
of the word and take all remaining characters that are still typed as
commands.
(Maybe create a PR in upstream?)