vim:keys: Fix <leader>grc when on end of hash
When the cursor is on the last character of the hash (e.g. after
pasting), the mapping would break and insert the commit subject behind
the next word.

Fix this by using `viw<Esc>` which places the cursor on the end of the
current word independent of the placement of the cursor.