vim:aucmd: Underline old cursor position
Underline old cursor position for better readability. Reusing
`TermCursor` made me confuse the current window many times.

My favorite solution would be an empty box around the position similar
to how terminals do it when focus is lost. But I fear that this is not
possible with a simple highlighting group.