zsh:alias: Manually expand aliases anywhere
When passing a full command as arguments to another (e.g. `sudo`),
aliases are not expanded.

Add the `$` alias that makes it possible to expand a following alias
anywhere. This uses the feature that aliases ending in a space will
trigger alias expansion on the next argument.

See: https://unix.stackexchange.com/a/433849