zsh:alias: Disable CORRECT_ALL for mkdir
It's really annoying if it tries to correct me when I create a new
directory that is named similar to an existing one.

Group together aliases that add a precommand modifier.

Since `mkdir` receives the `-p` flag already via `add_flags` the `md`
alias can directly alias to `mkdir` instead.