WIP: *: Try to conform to the REUSE specification
Add a SPDX license identifier and Copyright notice to more or less every
file and move the LICENSE into LICENSES/ to conform to the REUSE
specification (at the time of writing v3.3).

I decided against the (recommended) SPDX-FileCopyrightText tag for the
Copyright notice and instead opted for "Copyright (c)", because it:

	1. is allowed by the spec,
	2. is way more common (I have never seen the tag before)
	3. and in my opinion simply looks better (since this is to be
	   read by a human, I want the most readable option)

This change was primarily triggered by the upcoming import of
.clang-format from the Linux kernel, which includes already a SPDX
identifier and is licensed under GPL-2.0.

Link: https://reuse.software/