dircolors: Use a non-system color for DIR
I'm not a big fan of the standard blue that kitty ships with. Instead of
switching it's colors to my usual choice of Tango Dark, I want to adjust
the colors where they are used to theoretically make the terminal
experience a bit more terminal-agnostic.

SteelBlue3 (68) looks pretty to close to the blue I'm used to.